Dynamic Horizontal Scroll Bars
My god that was annoying to do...

I figure if you want to look for like 5 restaurants and then think about
it you can. But you dont want the page to be huge because of it. The
category sections are now scrollable horizontally.

Also, I've made a new class to identify our suggestions versus places
the person searches for.